Fix integer cast warnings for cast_unittest
Disable files that causes warnings for cast_unittest. Also fixed some
integer cast warnings.

R=justinlin@chromium.org

Review URL: https://codereview.chromium.org/25944002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226871 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/rtp_sender/rtp_packetizer/rtp_packetizer.h
@@ -65,7 +65,7 @@ class RtpPacketizer {
   uint16 packet_id_;
 
   int send_packets_count_;
-  int send_octet_count_;
+  size_t send_octet_count_;
 };
 
 }  // namespace cast
